+++ src/main/java/blackjack/view/InputView.java
@@ -1,4 +1,4 @@
-package blackjack.domain.view;
+package blackjack.view;
 
 import java.util.Scanner;
